Refactor readXxxFile
