+++ src/validate_policy.ts
@@ -7,8 +7,9 @@ const StreamUtils = require('@tilfin/stream-utils')
 import { listJsonFiles } from './utils/file'
 import { promisedStream } from './utils/stream'
 import { createWriter } from './utils/result_writer'
-import { readPolicyFile, PolicyEntry } from './aws/policy'
+import { PolicyEntry } from './aws/policy'
 import { PolicyValidator } from './logic/policy_validator'
+import { readPolicyFile } from './aws/file_reader';
 
 
 export async function main(inDir: string, varSet: any, opts: any = {}) {
